# Mat4 inverse and determinant SSE