Implements Mat4::inverse and Mat4::determinant using SIMD SSE intrinsics, based on GLM implementation.

Before:
```
PS C:\Users\Deprilola\ArquivosRapidos\glam-rs> cargo bench --bench mat4 "mat4 inverse"    
    Finished bench [optimized] target(s) in 0.14s
     Running target\release\deps\mat4-e24b0f8794d0c001.exe
Gnuplot not found, using plotters backend
mat4 inverse            time:   [22.059 ns 22.110 ns 22.163 ns]
                        change: [-0.8694% -0.3133% +0.2130%] (p = 0.26 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

PS C:\Users\Deprilola\ArquivosRapidos\glam-rs> cargo bench --bench mat4 "mat4 determinant"
    Finished bench [optimized] target(s) in 0.13s
     Running target\release\deps\mat4-e24b0f8794d0c001.exe
Gnuplot not found, using plotters backend
mat4 determinant        time:   [7.9098 ns 7.9423 ns 7.9773 ns]
                        change: [-0.4207% +0.3219% +1.1164%] (p = 0.42 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
```

After:
```
PS C:\Users\Deprilola\ArquivosRapidos\glam-rs> cargo bench --bench mat4 "mat4 inverse"
   Compiling glam v0.9.4 (C:\Users\Deprilola\ArquivosRapidos\glam-rs)
    Finished bench [optimized] target(s) in 5.08s
     Running target\release\deps\mat4-e24b0f8794d0c001.exe
Gnuplot not found, using plotters backend
mat4 inverse            time:   [15.380 ns 15.412 ns 15.444 ns]
                        change: [-31.967% -30.818% -29.974%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

PS C:\Users\Deprilola\ArquivosRapidos\glam-rs> cargo bench --bench mat4 "mat4 determinant"
    Finished bench [optimized] target(s) in 0.13s
     Running target\release\deps\mat4-e24b0f8794d0c001.exe
Gnuplot not found, using plotters backend
mat4 determinant        time:   [5.4185 ns 5.4306 ns 5.4438 ns]
                        change: [-35.908% -34.628% -33.449%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
```

Tests on my machine (Windows 10, i5-7400 @ 3GHz, 16GB 2666MHz)